Moved clip() from operations to rounding
