+++ heat/core/operations.py
@@ -18,7 +18,6 @@ __all__ = [
     'any',
     'argmax',
     'argmin',
-    'clip',
     'copy',
     'transpose',
     'tril',
@@ -367,41 +366,6 @@ def argmin(x, axis=None, out=None, **kwargs):
     return reduced_result
 
 
-def clip(a, a_min, a_max, out=None):
-    """
-    Parameters
-    ----------
-    a : ht.Tensor
-        Array containing elements to clip.
-    a_min : scalar or None
-        Minimum value. If None, clipping is not performed on lower interval edge. Not more than one of a_min and
-        a_max may be None.
-    a_max : scalar or None
-        Maximum value. If None, clipping is not performed on upper interval edge. Not more than one of a_min and
-        a_max may be None.
-    out : ht.Tensor, optional
-        The results will be placed in this array. It may be the input array for in-place clipping. out must be of
-        the right shape to hold the output. Its type is preserved.
-
-    Returns
-    -------
-    clipped_values : ht.Tensor
-        A tensor with the elements of this tensor, but where values < a_min are replaced with a_min, and those >
-        a_max with a_max.
-    """
-    if not isinstance(a, tensor.Tensor):
-        raise TypeError('a must be a tensor')
-    if a_min is None and a_max is None:
-        raise ValueError('either a_min or a_max must be set')
-
-    if out is None:
-        return tensor.Tensor(a._Tensor__array.clamp(a_min, a_max), a.shape, a.dtype, a.split, a.device, a.comm)
-    if not isinstance(out, tensor.Tensor):
-        raise TypeError('out must be a tensor')
-
-    return a._Tensor__array.clamp(a_min, a_max, out=out._Tensor__array) and out
-
-
 def copy(a):
     """
     Return an array copy of the given object.
@@ -652,7 +616,7 @@ def __local_operation(operation, x, out):
     # output buffer writing requires a bit more work
     # we need to determine whether the operands are broadcastable and the multiple of the broadcasting
     # reason: manually repetition for each dimension as PyTorch does not conform to numpy's broadcast semantic
-    # PyTorch always recreates the input shape and ignores broadcasting/too large buffers
+    # PyTorch always recreates the input shape and ignores broadcasting for too large buffers
     broadcast_shape = stride_tricks.broadcast_shape(x.lshape, out.lshape)
     padded_shape = (1,) * (len(broadcast_shape) - len(x.lshape)) + x.lshape
     multiples = [int(a / b) for a, b in zip(broadcast_shape, padded_shape)]
